Fix admin useCase

+++ domain/src/main/java/com/rcore/domain/user/usecase/admin/DeleteUserUseCase.java
@@ -1,18 +1,21 @@
 package com.rcore.domain.user.usecase.admin;
 
+import com.rcore.domain.token.exception.AuthenticationException;
 import com.rcore.domain.token.exception.AuthorizationException;
+import com.rcore.domain.token.usecase.AuthorizationByTokenUseCase;
 import com.rcore.domain.user.entity.UserEntity;
 import com.rcore.domain.user.exception.UserAlreadyExistException;
 import com.rcore.domain.user.port.UserRepository;
 import com.rcore.domain.user.access.AdminUserDeleteAccess;
 
-public class DeleteUserUseCase  extends AdminBaseUseCase {
+public class DeleteUserUseCase extends AdminBaseUseCase {
 
-    public DeleteUserUseCase(UserEntity actor, UserRepository userRepository) throws AuthorizationException {
-        super(actor, userRepository, new AdminUserDeleteAccess());
+    public DeleteUserUseCase(UserRepository userRepository, AuthorizationByTokenUseCase authorizationByTokenUseCase) {
+        super(userRepository, new AdminUserDeleteAccess(), authorizationByTokenUseCase);
     }
 
-    public Boolean delete(UserEntity userEntity) throws UserAlreadyExistException {
+    public Boolean delete(UserEntity userEntity) throws AuthenticationException, AuthorizationException {
+        checkAccess();
         return userRepository.delete(userEntity);
     }
