+++ domain/src/main/java/com/rcore/domain/userEmailVerification/usecase/admin/UserEmailVerificationDeleteUseCase.java
@@ -1,18 +1,22 @@
 package com.rcore.domain.userEmailVerification.usecase.admin;
 
+import com.rcore.domain.token.exception.AuthenticationException;
+import com.rcore.domain.token.usecase.AuthorizationByTokenUseCase;
 import com.rcore.domain.userEmailVerification.entity.UserEmailVerificationEntity;
 import com.rcore.domain.userEmailVerification.port.UserEmailVerificationRepository;
 import com.rcore.domain.userEmailVerification.access.AdminUserEmailVerificationDeleteAccess;
 import com.rcore.domain.token.exception.AuthorizationException;
 import com.rcore.domain.user.entity.UserEntity;
 
-public class UserEmailVerificationDeleteUseCase  extends UserEmailVerificationAdminBaseUseCase {
+public class UserEmailVerificationDeleteUseCase extends UserEmailVerificationAdminBaseUseCase {
 
-    public UserEmailVerificationDeleteUseCase(UserEntity actor, UserEmailVerificationRepository userEmailVerificationRepository) throws AuthorizationException {
-        super(actor, userEmailVerificationRepository, new AdminUserEmailVerificationDeleteAccess());
+    public UserEmailVerificationDeleteUseCase(UserEmailVerificationRepository userEmailVerificationRepository, AuthorizationByTokenUseCase authorizationByTokenUseCase) throws AuthorizationException {
+        super(userEmailVerificationRepository, new AdminUserEmailVerificationDeleteAccess(), authorizationByTokenUseCase);
     }
 
-    public Boolean delete(UserEmailVerificationEntity userEmailVerificationEntity){
+    public Boolean delete(UserEmailVerificationEntity userEmailVerificationEntity) throws AuthenticationException, AuthorizationException {
+        checkAccess();
+
         userEmailVerificationRepository.delete(userEmailVerificationEntity);
 
         return true;
