+++ domain/src/main/java/com/rcore/domain/userLog/usecase/admin/UserLogDeleteUseCase.java
@@ -1,5 +1,7 @@
 package com.rcore.domain.userLog.usecase.admin;
 
+import com.rcore.domain.token.exception.AuthenticationException;
+import com.rcore.domain.token.usecase.AuthorizationByTokenUseCase;
 import com.rcore.domain.userLog.entity.UserLogEntity;
 import com.rcore.domain.userLog.port.UserLogRepository;
 import com.rcore.domain.userLog.access.AdminUserLogDeleteAccess;
@@ -8,11 +10,12 @@ import com.rcore.domain.user.entity.UserEntity;
 
 public class UserLogDeleteUseCase  extends UserLogAdminBaseUseCase {
 
-    public UserLogDeleteUseCase(UserEntity actor, UserLogRepository userLogRepository) throws AuthorizationException {
-        super(actor, userLogRepository, new AdminUserLogDeleteAccess());
+    public UserLogDeleteUseCase(UserLogRepository userLogRepository, AuthorizationByTokenUseCase authorizationByTokenUseCase){
+        super(userLogRepository, new AdminUserLogDeleteAccess(), authorizationByTokenUseCase);
     }
 
-    public Boolean delete(UserLogEntity userLogEntity){
+    public Boolean delete(UserLogEntity userLogEntity) throws AuthenticationException, AuthorizationException {
+        checkAccess();
         userLogRepository.delete(userLogEntity);
 
         return true;
